refactor(tree-tests): Separate related tree node tests into smaller files
+ Separate related tree node tests into smaller files
+ Extract common stub values to `RegexTreeStubs`
